requirements.txt
taichi
numpy-quaternion